chore: consolidate HMAC platform splits using Web Crypto
Replace node:crypto HMAC implementations with Web Crypto API (available
in Node 20+, browser, and React Native). Delete browser/RN shims that
just re-exported the common Web Crypto implementation.

- core-amqp: collapse hmacSha256.ts + common + web shim → single file
- service-bus: collapse crypto.ts + common + browser + RN → single file
- notification-hubs: collapse hmacSha256.ts + common + browser + RN → single file

Use uint8ArrayToString from @azure/core-util for base64 encoding instead
of btoa(String.fromCharCode(...)) hack.
