--- sdk/core/core-amqp/src/util/hmacSha256-web.mts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export * from "./hmacSha256.common.js";
