--- sdk/notificationhubs/notification-hubs/src/auth/hmacSha256-react-native.mts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export { signString } from "./hmacSha256.common.js";
