+++ sdk/notificationhubs/notification-hubs/src/auth/hmacSha256.ts
@@ -1,9 +1,25 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { createHmac } from "node:crypto";
+import { uint8ArrayToString } from "@azure/core-util";
 
 export async function signString(key: string, toSign: string): Promise<string> {
-  const hmac = createHmac("sha256", key).update(toSign).digest("base64");
-  return encodeURIComponent(hmac);
+  const enc = new TextEncoder();
+  const algorithm = { name: "HMAC", hash: { name: "SHA-256" } };
+
+  const extractedKey = await globalThis.crypto.subtle.importKey(
+    "raw",
+    enc.encode(key),
+    algorithm,
+    false,
+    ["sign", "verify"],
+  );
+  const signature = await globalThis.crypto.subtle.sign(
+    algorithm,
+    extractedKey,
+    enc.encode(toSign),
+  );
+  const digest = uint8ArrayToString(new Uint8Array(signature), "base64");
+
+  return encodeURIComponent(digest);
 }
