--- sdk/servicebus/service-bus/src/util/crypto-browser.mts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export { generateKey } from "./crypto.common.js";
