--- sdk/servicebus/service-bus/src/util/crypto.common.ts
@@ -1,52 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-/**
- * @internal
- */
-export async function generateKey(secret: string, stringToSign: string): Promise<string> {
-  const key = await globalThis.crypto.subtle.importKey(
-    "raw",
-    convertToUint8Array(secret),
-    {
-      name: "HMAC",
-      hash: { name: "SHA-256" },
-    },
-    false,
-    ["sign"],
-  );
-
-  const signature = await globalThis.crypto.subtle.sign(
-    "HMAC",
-    key,
-    convertToUint8Array(stringToSign),
-  );
-  const base64encodedString = encodeByteArray(new Uint8Array(signature));
-  const result = encodeURIComponent(base64encodedString);
-  return result;
-}
-
-/**
- * @internal
- */
-function convertToUint8Array(value: string): Uint8Array<ArrayBuffer> {
-  const arr = new Uint8Array(value.length);
-  for (let i = 0; i < value.length; i++) {
-    arr[i] = value.charCodeAt(i);
-  }
-  return arr;
-}
-
-/**
- * Encodes a byte array in base64 format.
- * @param value - the Uint8Array to encode
- * @internal
- *
- */
-function encodeByteArray(value: Uint8Array): string {
-  let str = "";
-  for (let i = 0; i < value.length; i++) {
-    str += String.fromCharCode(value[i]);
-  }
-  return btoa(str);
-}
